+++ serial/storage.h
@@ -21,6 +21,6 @@
  *
  */
 
-int port_delete(bdaddr_t *src, bdaddr_t *dst, int id);
-int port_store(bdaddr_t *src, bdaddr_t *dst, int id,
+int port_delete(bdaddr_t *src, bdaddr_t *dst, int16_t id);
+int port_store(bdaddr_t *src, bdaddr_t *dst, int16_t id,
 			uint8_t ch, const char *svcname);
